Some fixes

+++ rater/views.py
@@ -83,8 +83,9 @@ def random_meme(request):
     if request.method == "POST":
         meme_img = MemeImage.objects.filter(id=request.POST.get("pk", 0)).first()
         if int(request.POST.get("quality", 1)) == 0:
-            meme_img.delete()
-            print("Deleted meme.")
+            if meme_img:
+                meme_img.delete()
+                print("Deleted meme.")
 
     meme = random.choice(Meme.objects.all())
